move inactivate to separate file
